+++ packages/clouds/src/CloudsEffect.ts
@@ -707,20 +707,40 @@ export class CloudsEffect extends Effect {
     this.cloudsPass.currentMaterial.scatterAnisotropyMix = value
   }
 
+  /** @deprecated Use skyLightScale instead. */
   get skyIrradianceScale(): number {
-    return this.cloudsPass.currentMaterial.uniforms.skyIrradianceScale.value
+    return this.skyLightScale
   }
 
+  /** @deprecated Use skyLightScale instead. */
   set skyIrradianceScale(value: number) {
-    this.cloudsPass.currentMaterial.uniforms.skyIrradianceScale.value = value
+    this.skyLightScale = value
   }
 
+  get skyLightScale(): number {
+    return this.cloudsPass.currentMaterial.uniforms.skyLightScale.value
+  }
+
+  set skyLightScale(value: number) {
+    this.cloudsPass.currentMaterial.uniforms.skyLightScale.value = value
+  }
+
+  /** @deprecated Use groundBounceScale instead. */
   get groundIrradianceScale(): number {
-    return this.cloudsPass.currentMaterial.uniforms.groundIrradianceScale.value
+    return this.groundBounceScale
   }
 
+  /** @deprecated Use groundBounceScale instead. */
   set groundIrradianceScale(value: number) {
-    this.cloudsPass.currentMaterial.uniforms.groundIrradianceScale.value = value
+    this.groundBounceScale = value
+  }
+
+  get groundBounceScale(): number {
+    return this.cloudsPass.currentMaterial.uniforms.groundBounceScale.value
+  }
+
+  set groundBounceScale(value: number) {
+    this.cloudsPass.currentMaterial.uniforms.groundBounceScale.value = value
   }
 
   get powderScale(): number {